# Reimplement IPSet.Ranges with a (somewhat) easier to follow algorithm.