```
    The main goal was to fix #125 by providing a less tricky implementation
    of Ranges as a comparison baseline... But this implementation turns out
    to also be a bit faster and more memory efficient.
    
    Fixes #125.
    
    name         old time/op    new time/op    delta
    IPSetFuzz-8    16.3µs ± 3%    14.1µs ± 1%  -12.99%  (p=0.008 n=5+5)
    
    name         old alloc/op   new alloc/op   delta
    IPSetFuzz-8    2.60kB ± 0%    1.95kB ± 0%  -25.24%  (p=0.008 n=5+5)
    
    name         old allocs/op  new allocs/op  delta
    IPSetFuzz-8      33.0 ± 0%      30.0 ± 0%   -9.09%  (p=0.008 n=5+5)
    
    Signed-off-by: David Anderson <dave@natulte.net>
```