@danderson 2021-01-27
Ready for review @bradfitz @mdlayher. Tests pass at all commits (I only add the previously-failing test case after the major commit that fixes the bug), so can be rebase-and-merged.