@danderson 2021-01-29
Looks much better with the symbolic names for each case, thanks. Applied all feedback, going to submit and we can bikeshed on the exact names of helpers in a followup (I'm not _super_ happy with some of the subtler ones).